Add Power panel, chart updates, and UI color changes
Updated XAML to add Power (W) panel and use distinct colors for metric borders. Added Power property to PowerReading and displayed it in the UI. Changed ChartData.Size to float and implemented live chart updates with time tracking. Refactored Bluetooth alerts and removed toggle switch logic. Improved async event handling.

+++ ESP32MAUIPowerMonitoringBLE/ChartData.cs
@@ -7,7 +7,7 @@ namespace ESP32MAUIPowerMonitoringBLE
     public class ChartData
     {
         public double Value { get; set; } // x-axis (time)
-        public double Size { get; set; }  // y-axis (power)
+        public float Size { get; set; }  // y-axis (power)
     }
 
 }
